requirements.txt
streamlit
pypdf
langchain
langchain-google-genai
langchain-groq
langchain-text-splitters
python-dotenv
